+++ src/ChronicleServiceProvider.php
@@ -16,6 +16,7 @@ use Chronicle\Console\Commands\KeyRotateCommand;
 use Chronicle\Console\Commands\LegalHoldCommand;
 use Chronicle\Console\Commands\PruneCommand;
 use Chronicle\Console\Commands\ReportCommand;
+use Chronicle\Console\Commands\RotateKekCommand;
 use Chronicle\Console\Commands\ShowEntryCommand;
 use Chronicle\Console\Commands\StatsCommand;
 use Chronicle\Console\Commands\SubjectEraseCommand;
@@ -117,6 +118,7 @@ class ChronicleServiceProvider extends ServiceProvider
                 LegalHoldCommand::class,
                 SubjectEraseCommand::class,
                 SubjectKeysCommand::class,
+                RotateKekCommand::class,
             ]);
         }
     }